minor ts issue
