+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+  "spellright.language": ["en"],
+  "spellright.documentTypes": ["markdown", "latex", "plaintext"]
+}
